chore: warning moved to pino.final
