+++ lib/tools.js
@@ -437,6 +437,9 @@ function createArgsNormalizer (defaultOptions) {
 }
 
 function final (logger, handler) {
+  const major = Number(process.versions.node.split('.')[0])
+  if (major >= 14) warning.emit('PINODEP009')
+
   if (typeof logger === 'undefined' || typeof logger.child !== 'function') {
     throw Error('expected a pino logger instance')
   }
